+++ compact/models.go
@@ -19,7 +19,7 @@ type CompactForms map[CompactType][]CompactFormRule
 // ZeroesInPattern indicates the number of zero characters in the default "other" pattern, and PatternsByPluralForm
 // is a map of plural forms (e.g. one, few, many, other, etc.) to patterns.
 type CompactFormRule struct {
-	Type                 int
+	Type                 uint64
 	ZeroesInPattern      int
 	PatternsByPluralForm map[string]string
 }
